+++ pom.xml
@@ -124,6 +124,13 @@
             <artifactId>gson</artifactId>
             <version>2.3.1</version>
         </dependency>
+        <dependency>
+            <groupId>com.github.fge</groupId>
+            <artifactId>jackson-coreutils</artifactId>
+            <version>1.8</version>
+            <type>jar</type>
+        </dependency>
+
         
         <dependency>
             <groupId>org.yaml</groupId>
